+++ src/entities/user_details.rs
@@ -1,30 +1,165 @@
 use comfy_table::{ Table, ContentArrangement };
 
-#[derive(Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct UserDetails {
     #[serde(rename = "Name")]
     pub name: String,
     #[serde(rename = "ServerId")]
-    serverid: String,
+    pub server_id: String,
+    #[serde(rename = "ServerName", default)]
+    pub server_name: String,
     #[serde(rename = "Id")]
     pub id: String,
+    #[serde(rename = "PrimaryImageTag", default)]
+    pub primary_image_tag: String,
+    #[serde(rename = "HasPassword")]
+    pub has_password: bool,
+    #[serde(rename = "HasConfiguredPassword")]
+    pub has_configured_password: bool,
+    #[serde(rename = "HasConfiguredEasyPassword")]
+    pub has_configured_easy_password: bool,
+    #[serde(rename = "EnableAutoLogin")]
+    pub enable_auto_login: bool,
+    #[serde(rename = "LastLoginDate", default)]
+    pub last_login_date: Option<String>,
+    #[serde(rename = "LastActivityDate", default)]
+    pub last_activity_date: Option<String>,
+    #[serde(rename = "Configuration")]
+    pub configuration: Configuration,
     #[serde(rename = "Policy")]
-    pub policy: Policy
-}  
-
-// Struct to contain the Policy information that is a part of the user details.
-#[derive(Serialize, Deserialize)]
-pub struct Policy {
-    #[serde(rename = "AuthenticationProviderId")]
-    pub auth_provider_id: String,
-    #[serde(rename = "PasswordResetProviderId")]
-    pub pass_reset_provider_id: String,
-    #[serde(rename = "IsAdministrator")]
-    is_admin: bool,
-    #[serde(rename = "IsDisabled")]
-    is_disabled: bool
+    pub policy: Policy,
+    #[serde(rename = "PrimaryImageAspectRatio", default)]
+    pub primary_image_aspect_ratio: i64,
 }
 
+// Struct to contain the Policy information that is a part of the user details.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Policy {
+    #[serde(rename = "IsAdministrator")]
+    pub is_administrator: bool,
+    #[serde(rename = "IsHidden")]
+    pub is_hidden: bool,
+    #[serde(rename = "IsDisabled")]
+    pub is_disabled: bool,
+    #[serde(rename = "MaxParentalRating", default)]
+    pub max_parental_rating: i64,
+    #[serde(rename = "BlockedTags")]
+    pub blocked_tags: Vec<String>,
+    #[serde(rename = "EnableUserPreferenceAccess")]
+    pub enable_user_preference_access: bool,
+    #[serde(rename = "AccessSchedules")]
+    pub access_schedules: Vec<AccessSchedule>,
+    #[serde(rename = "BlockUnratedItems")]
+    pub block_unrated_items: Vec<String>,
+    #[serde(rename = "EnableRemoteControlOfOtherUsers")]
+    pub enable_remote_control_of_other_users: bool,
+    #[serde(rename = "EnableSharedDeviceControl")]
+    pub enable_shared_device_control: bool,
+    #[serde(rename = "EnableRemoteAccess")]
+    pub enable_remote_access: bool,
+    #[serde(rename = "EnableLiveTvManagement")]
+    pub enable_live_tv_management: bool,
+    #[serde(rename = "EnableLiveTvAccess")]
+    pub enable_live_tv_access: bool,
+    #[serde(rename = "EnableMediaPlayback")]
+    pub enable_media_playback: bool,
+    #[serde(rename = "EnableAudioPlaybackTranscoding")]
+    pub enable_audio_playback_transcoding: bool,
+    #[serde(rename = "EnableVideoPlaybackTranscoding")]
+    pub enable_video_playback_transcoding: bool,
+    #[serde(rename = "EnablePlaybackRemuxing")]
+    pub enable_playback_remuxing: bool,
+    #[serde(rename = "ForceRemoteSourceTranscoding")]
+    pub force_remote_source_transcoding: bool,
+    #[serde(rename = "EnableContentDeletion")]
+    pub enable_content_deletion: bool,
+    #[serde(rename = "EnableContentDeletionFromFolders")]
+    pub enable_content_deletion_from_folders: Vec<String>,
+    #[serde(rename = "EnableContentDownloading")]
+    pub enable_content_downloading: bool,
+    #[serde(rename = "EnableSyncTranscoding")]
+    pub enable_sync_transcoding: bool,
+    #[serde(rename = "EnableMediaConversion")]
+    pub enable_media_conversion: bool,
+    #[serde(rename = "EnabledDevices")]
+    pub enabled_devices: Vec<String>,
+    #[serde(rename = "EnableAllDevices")]
+    pub enable_all_devices: bool,
+    #[serde(rename = "EnabledChannels")]
+    pub enabled_channels: Vec<String>,
+    #[serde(rename = "EnableAllChannels")]
+    pub enable_all_channels: bool,
+    #[serde(rename = "EnabledFolders")]
+    pub enabled_folders: Vec<String>,
+    #[serde(rename = "EnableAllFolders")]
+    pub enable_all_folders: bool,
+    #[serde(rename = "InvalidLoginAttemptCount")]
+    pub invalid_login_attempt_count: i64,
+    #[serde(rename = "LoginAttemptsBeforeLockout")]
+    pub login_attempts_before_lockout: i64,
+    #[serde(rename = "MaxActiveSessions")]
+    pub max_active_sessions: i64,
+    #[serde(rename = "EnablePublicSharing")]
+    pub enable_public_sharing: bool,
+    #[serde(rename = "BlockedMediaFolders")]
+    pub blocked_media_folders: Vec<String>,
+    #[serde(rename = "BlockedChannels")]
+    pub blocked_channels: Vec<String>,
+    #[serde(rename = "RemoteClientBitrateLimit")]
+    pub remote_client_bitrate_limit: i64,
+    #[serde(rename = "AuthenticationProviderId")]
+    pub authentication_provider_id: String,
+    #[serde(rename = "PasswordResetProviderId")]
+    pub password_reset_provider_id: String,
+    #[serde(rename = "SyncPlayAccess")]
+    pub sync_play_access: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Configuration {
+    #[serde(rename = "AudioLanguagePreference", default)]
+    pub audio_language_preference: String,
+    #[serde(rename = "PlayDefaultAudioTrack")]
+    pub play_default_audio_track: bool,
+    #[serde(rename = "SubtitleLanguagePreference")]
+    pub subtitle_language_preference: String,
+    #[serde(rename = "DisplayMissingEpisodes")]
+    pub display_missing_episodes: bool,
+    #[serde(rename = "GroupedFolders")]
+    pub grouped_folders: Vec<String>,
+    #[serde(rename = "SubtitleMode")]
+    pub subtitle_mode: String,
+    #[serde(rename = "DisplayCollectionsView")]
+    pub display_collections_view: bool,
+    #[serde(rename = "EnableLocalPassword")]
+    pub enable_local_password: bool,
+    #[serde(rename = "OrderedViews")]
+    pub ordered_views: Vec<String>,
+    #[serde(rename = "LatestItemsExcludes")]
+    pub latest_items_excludes: Vec<String>,
+    #[serde(rename = "MyMediaExcludes")]
+    pub my_media_excludes: Vec<String>,
+    #[serde(rename = "HidePlayedInLatest")]
+    pub hide_played_in_latest: bool,
+    #[serde(rename = "RememberAudioSelections")]
+    pub remember_audio_selections: bool,
+    #[serde(rename = "RememberSubtitleSelections")]
+    pub remember_subtitle_selections: bool,
+    #[serde(rename = "EnableNextEpisodeAutoPlay")]
+    pub enable_next_episode_auto_play: bool,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct AccessSchedule {
+    #[serde(rename = "UserId")]
+    pub user_id: String,
+    #[serde(rename = "DayOfWeek")]
+    pub day_of_week: String,
+    #[serde(rename = "StartHour")]
+    pub start_hour: i64,
+    #[serde(rename = "EndHour")]
+    pub end_hour: i64,
+}
 impl UserDetails {
     pub fn json_print(users: Vec<UserDetails>) {
         println!("{}", serde_json::to_string_pretty(&users).unwrap());
@@ -37,7 +172,7 @@ impl UserDetails {
             .set_width(120)
             .set_header(vec!["Username", "Admin", "Disabled"]);
         for user in users {
-            table.add_row(vec![user.name, user.policy.is_admin.to_string(), user.policy.is_disabled.to_string()]);
+            table.add_row(vec![user.name, user.policy.is_administrator.to_string(), user.policy.is_disabled.to_string()]);
         }
         println!("{table}");
     }